переносим все в один класс (причина: мало кода для разделения в два класса)
