+++ FlameEnchant/src/Pechenka/flame/FlameEnchant.php
@@ -10,14 +10,19 @@ namespace Pechenka\flame;
  *    \/    \___|\___|_| |_|\___|_| |_|_|\_\__,_|
  */
 
+use pocketmine\event\entity\EntityDamageByEntityEvent;
+use pocketmine\event\entity\EntityDamageEvent;
+use pocketmine\event\entity\EntityShootBowEvent;
+use pocketmine\event\Listener;
+use pocketmine\item\enchantment\Enchantment;
+use pocketmine\Player;
 use pocketmine\plugin\PluginBase;
-use Pechenka\flame\events\EventListener;
 
 /**
  * Class FlameEnchant
  * @package Pechenka\flame
  */
-class FlameEnchant extends PluginBase {
+class FlameEnchant extends PluginBase implements Listener{
 
     /**
      * @var FlameEnchant
@@ -27,7 +32,7 @@ class FlameEnchant extends PluginBase {
 
     public function onEnable() {
         self::$instance = $this;
-        $this->getServer()->getPluginManager()->registerEvents(new EventListener($this), $this);
+        $this->getServer()->getPluginManager()->registerEvents($this, $this);
         $this->getLogger()->info("Плагин FlameEnchant был успешно включен! Автор - 3P: vk.com/vovan446");
     }
 
@@ -36,5 +41,35 @@ class FlameEnchant extends PluginBase {
      */
     public static function getPlugin() { //ненужная функция, можете убрать ее, ошибок не будет, возможно кто-то захочет допилить зачарования и ему она пригодится.
         return self::$instance;
+    }
+
+    /**
+     * @param EntityDamageEvent $event
+     * @ignoreCancelled false
+     */
+    public function swordFlame(EntityDamageEvent $event) {
+        $player = $event->getEntity();
+        if ($player instanceof Player) {
+            if ($event instanceof EntityDamageByEntityEvent) {
+                $damager = $event->getDamager();
+                if ($damager instanceof Player) {
+                    $item = $damager->getInventory()->getItemInHand();
+                    if ($item->hasEnchantment(Enchantment::FIRE_ASPECT)) {
+                        $player->setOnFire(3);
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * @param EntityShootBowEvent $event
+     * @ignoreCancelled false
+     */
+    public function bowFlame(EntityShootBowEvent $event) {
+        $entity = $event->getProjectile();
+        if (($bow = $event->getBow()) !== null && $bow->hasEnchantment(Enchantment::FLAME)) {
+            $entity->setOnFire(777);
+        }
     }
 }
