--- FlameEnchant/src/Pechenka/flame/events/EventListener.php
@@ -1,67 +1,0 @@
-<?php
-
-namespace Pechenka\flame\events;
-
-/**
- *       ___          _                _
- *     / _ \___  ___| |__   ___ _ __ | | ____ _
- *    / /_)/ _ \/ __| '_ \ / _ \ '_ \| |/ / _` |
- *    / ___/  __/ (__| | | |  __/ | | |   < (_| |
- *    \/    \___|\___|_| |_|\___|_| |_|_|\_\__,_|
- */
-
-use pocketmine\event\Listener;
-use pocketmine\Player;
-use pocketmine\item\Enchantment;
-use pocketmine\event\entity\EntityDamageEvent;
-use pocketmine\event\entity\EntityDamageByEntityEvent;
-use pocketmine\event\entity\EntityShootBowEvent;
-use Pechenka\flame\FlameEnchant;
-
-/**
- * Class EventListener
- * @package Pechenka\flame\events
- */
-class EventListener implements Listener {
-
-
-    /**
-     * EventListener constructor.
-     * @param FlameEnchant $plugin
-     */
-    public function __construct(FlameEnchant $plugin) {
-        self::$instance = $plugin;
-    }
-
-
-
-    /**
-     * @param EntityDamageEvent $event
-     * @ignoreCancelled false
-     */
-    public function swordFlame(EntityDamageEvent $event) {
-        $player = $event->getEntity();
-        if ($player instanceof Player) {
-            if ($event instanceof EntityDamageByEntityEvent) {
-                $damager = $event->getDamager();
-                if ($damager instanceof Player) {
-                    $item = $damager->getInventory()->getItemInHand();
-                    if ($item->hasEnchantment(Enchantment::FIRE_ASPECT)) {    
-                        $player->setOnFire(3);
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * @param EntityShootBowEvent $event
-     * @ignoreCancelled false
-     */
-    public function bowFlame(EntityShootBowEvent $event) {
-        $entity = $event->getProjectile();
-        if (($bow = $event->getBow()) !== null && $bow->hasEnchantment(Enchantment::FLAME)) { 
-            $entity->setOnFire(777);
-        }
-    }
-}
